Move settings label building into a mapper
